Convert to Go Modules

+++ dist.sh
@@ -6,15 +6,15 @@ set -e
 DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 echo "working dir $DIR"
 
-rm -rf $DIR/vendor
-echo "... refreshing vendor directory"
-./vendor.sh
+if [ -e vendor ]; then
+	rm -rf $DIR/vendor
+fi
 
 echo "... running tests"
-gb test || exit 1
+go test ./... || exit 1
 
 arch=$(go env GOARCH)
-version=$(cat $DIR/src/cmd/git-open-pull/version.go | grep "const Version" | awk '{print $NF}' | sed 's/"//g')
+version=$(cat $DIR/version.go | grep "const Version" | awk '{print $NF}' | sed 's/"//g')
 goversion=$(go version | awk '{print $3}')
 
 mkdir -p dist
@@ -22,9 +22,8 @@ for os in linux darwin; do
     echo "... building v$version for $os/$arch"
     BUILD=$(mktemp -d -t git-open-pull)
     TARGET="git-open-pull-$version.$os-$arch.$goversion"
-    GOOS=$os GOARCH=$arch CGO_ENABLED=0 gb build
     mkdir -p $BUILD/$TARGET
-    cp bin/git-open-pull-$os-$arch $BUILD/$TARGET/git-open-pull
+    GOOS=$os GOARCH=$arch CGO_ENABLED=0 go build -o $BUILD/$TARGET/git-open-pull
     pushd $BUILD >/dev/null
     tar czvf $TARGET.tar.gz $TARGET
     if [ -e $DIR/dist/$TARGET.tar.gz ]; then
